Clean up files, add dogs api test
