+++ client/package.json
@@ -35,7 +35,6 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "http://localhost:5000",
   "devDependencies": {
     "jest-fetch-mock": "^3.0.1"
   }
