+++ client/src/data/dogs.js
@@ -1,5 +1,5 @@
 export default async function fetchDogs() {
-  const response = await fetch("http://localhost:3000/dogs");
+  const response = await fetch("http://localhost:5000/dogs");
   const json = await response.json();
   return json;
 }
